Extract out common generator code for macaroons
Given the handler and the SciTokens plugin will want to invoke the
macaroon generation code, extract it out into a standalone class in
the XrdMacaroons directory.

The intent of this code is to keep the libmacaroons code inside the
XrdMacaroons directory.  The cmake changes to build this code will
eventually be replaced with a plugin interface.

A subsequent commit will switch XrdMacaroons library to utilize
this generator code.
