+++ src/XrdMacaroons.cmake
@@ -15,6 +15,7 @@ if( BUILD_MACAROONS )
     XrdMacaroons/XrdMacaroons.cc
     XrdMacaroons/XrdMacaroonsHandler.cc     XrdMacaroons/XrdMacaroonsHandler.hh
     XrdMacaroons/XrdMacaroonsAuthz.cc       XrdMacaroons/XrdMacaroonsAuthz.hh
+    XrdMacaroons/XrdMacaroonsGenerate.cc    XrdMacaroons/XrdMacaroonsGenerate.hh
     XrdMacaroons/XrdMacaroonsUtils.cc       XrdMacaroons/XrdMacaroonsUtils.hh
     XrdMacaroons/XrdMacaroonsConfigure.cc)
