+++ src/XrdSciTokens.cmake
@@ -13,10 +13,11 @@ add_library(
    ${LIB_XRD_SCITOKENS}
    MODULE
    XrdSciTokens/XrdSciTokensAccess.cc
-   XrdMacaroons/XrdMacaroonsUtils.cc   XrdMacaroons/XrdMacaroonsUtils.hh
-                                       XrdSciTokens/XrdSciTokensRedir.hh
-                                       XrdSciTokens/XrdSciTokensHelper.hh
-   XrdSciTokens/XrdSciTokensMon.cc     XrdSciTokens/XrdSciTokensMon.hh )
+   XrdMacaroons/XrdMacaroonsUtils.cc    XrdMacaroons/XrdMacaroonsUtils.hh
+   XrdMacaroons/XrdMacaroonsGenerate.cc XrdMacaroons/XrdMacaroonsGenerate.hh
+                                        XrdSciTokens/XrdSciTokensRedir.hh
+                                        XrdSciTokens/XrdSciTokensHelper.hh
+   XrdSciTokens/XrdSciTokensMon.cc      XrdSciTokens/XrdSciTokensMon.hh )
 target_link_libraries(
    ${LIB_XRD_SCITOKENS}
    PRIVATE
